Extract NoisePayload into a separate file
